pyproject.toml
[project]
name = "resting-agent"
version = "0.1.0"
description = "Autonomous agent for building RESTful APIs"
readme = "README.md"
authors = [{ name = "ethan-wickstrom", email = "e.t.wickstrom@wustl.edu" }]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Code Generators",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "dspy-ai>=2.6.27",
    "pydantic>=2.11.7",
    "pyright>=1.1.402",
    "typer>=0.16.0",
]

[project.urls]
Homepage = "https://github.com/ethan-wickstrom/resting-agent"
Repository = "https://github.com/ethan-wickstrom/resting-agent.git"
Issues = "https://github.com/ethan-wickstrom/resting-agent/issues"
Documentation = "https://github.com/ethan-wickstrom/resting-agent#readme"

[project.scripts]
resting-agent = "resting_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resting_agent"]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "RUF", # ruff-specific rules
]
ignore = [
  "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # allow assert in tests

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "dspy.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=resting_agent",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/resting_agent"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.pyright]
venvPath = "."
venv = ".venv"
